supplementary_materials/qwen2/requirements.txt
nncf==2.10.0
optimum-intel>=1.16.0
transformers>=4.38.2
openvino>=2024.1.0
onnx>=1.15.0
huggingface-hub>=0.21.3
modelscope
